Revert "Merge branch 'en/rename-directory-detection'"
This reverts commit 447eda2e74cc9e4cb3578355d1e5d977ba904049, reversing
changes made to 51aa245879d26bd2421cd078c9581d2e7215254a.

The topic appears to inflict severe regression in renaming merges,
even though the promise of it was that it would improve them.

We do not yet know which exact change in the topic was wrong, but in
the meantime, let's play it safe and revert it out of 'master'
before real Git-using projects are harmed.

Signed-off-by: Junio C Hamano <gitster@pobox.com>